hashing: fix the implementation
The offset or increment on collision should be added to the start
position (initial hash value that resulted in collision).

Previously, the increment was being added to the initial hash value
and the result stored in a current position variable (pos).
The subsequent offset additions were being done to the position
variable, instead of the start position. This goes against the
logic.

Signed-off-by: ykhan <yasir_electronics@yahoo.com>